+++ src/background.js
@@ -1,0 +1,3 @@
+const openOptionsPage = () => chrome.runtime.openOptionsPage();
+
+(chrome.action ?? chrome.browserAction).onClicked.addListener(openOptionsPage);