+++ src/utils/talespire.js
@@ -1,3 +1,36 @@
+import { settings } from "~/utils/storage";
+
+const modifierAction = (action, name) => {
+  switch (action) {
+    case "adv":
+      return { name: `${name} (ADV)`, extraDice: true };
+    case "dis":
+      return { name: `${name} (DIS)`, extraDice: true };
+    case "adv-dis":
+      return { name: `${name} (ADV/DIS)`, extraDice: true };
+
+    default:
+      return { name, extraDice: false };
+  }
+};
+
+const checkModifierKeys = (event, name) => {
+  let action = "none";
+  if (event.altKey) {
+    action = settings.modifierKeyAlt;
+  } else if (event.ctrlKey) {
+    action = settings.modifierKeyCtrl;
+  } else if (event.shiftKey) {
+    action = settings.modifierKeyShift;
+  }
+
+  if (action === "none") {
+    return { name, extraDice: false };
+  }
+
+  return modifierAction(action, name);
+};
+
 export const talespireLink = (elem, label, dice, diceLabel) => {
   label = label?.trim();
 
@@ -11,18 +44,13 @@ export const talespireLink = (elem, label, dice, diceLabel) => {
 
     dice = dice.replace(/d100/g, "d100+d10");
 
-    let name = label;
-    let extraDice = "";
-    if (event.altKey || event.ctrlKey || event.shiftKey) {
-      name += " (ADV/DIS)";
-      extraDice = `/${dice}`;
-    }
+    const { name, extraDice } = checkModifierKeys(event, label);
 
     let uri;
     if (typeof name === "string") {
-      uri = `talespire://dice/${encodeURIComponent(name)}:${dice}${extraDice}`;
+      uri = `talespire://dice/${encodeURIComponent(name)}:${dice}${extraDice ? `/${dice}` : ""}`;
     } else {
-      uri = `talespire://dice/${dice}${extraDice}`;
+      uri = `talespire://dice/${dice}${extraDice ? `/${dice}` : ""}`;
     }
 
     if (TB_DRY_RUN_TALESPIRE_LINKS === "true") {